requirements.txt
Flask>=3.0.2
Jinja2>=3.1.3
chardet>=5.2.0
cssselect>=1.2.0
lxml>=5.1.0
pycurl>=7.45.3
pyquery>=2.0.0
requests>=2.31.0
tornado>=6.4.0
mysql-connector-python>=8.3.0
pika>=1.3.2
pymongo>=4.6.1
Flask-Login>=0.6.3
u-msgpack-python>=2.8.0
click>=8.1.7
SQLAlchemy>=2.0.28
six>=1.16.0
amqp>=5.2.0
redis>=5.0.1
kombu>=5.3.5
psycopg2>=2.9.9
elasticsearch>=8.12.1
tblib>=3.0.0
wsgidav>=4.3.0
# API dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0
dicttoxml>=1.7.0
PyYAML>=6.0.0
python-multipart>=0.0.6
